+++ PlayTV.php
@@ -245,23 +245,26 @@
     {
       var $headers, $user_agent, $compression, $cookie_file;
       var $cert_check, $proxy;
+      static $ref = 0;
 
       function cURL($cookies = true, $cookie = 'Cookies.txt', $compression = 'gzip', $proxy = '')
         {
           $this->headers     = $this->headers();
           $this->user_agent  = 'Mozilla/5.0 (Windows NT 5.1; rv:15.0) Gecko/20100101 Firefox/15.0';
           $this->compression = $compression;
-          $this->proxy       = $proxy;
           $this->cookies     = $cookies;
-          $this->cert_check  = true;
           if ($this->cookies == true)
               $this->cookie($cookie);
+          $this->cert_check = true;
+          $this->proxy      = $proxy;
+          self::$ref++;
         }
 
       function __destruct()
         {
-          if (file_exists($this->cookie_file))
+          if ((self::$ref <= 1) and file_exists($this->cookie_file))
               unlink($this->cookie_file);
+          self::$ref--;
         }
 
       function headers()
